moved services fixture to confitest
